+++ lib/controllers/notification_controller.dart
@@ -7,7 +7,6 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:mi_utem/config/routes.dart';
 import 'package:mi_utem/models/asignatura.dart';
-import 'package:mi_utem/screens/asignatura_detalle_screen.dart';
 import 'package:mi_utem/services/analytics_service.dart';
 
 class NotificationController {
@@ -56,10 +55,7 @@ class NotificationController {
         );
         final asignatura =
             Asignatura.fromJson(jsonDecode(asignaturaJsonString));
-        Get.to(
-          () => AsignaturaDetalleScreen(asignatura: asignatura),
-          routeName: Routes.asignatura,
-        );
+        Get.toNamed('${Routes.asignatura}/${asignatura.id}');
       }
     }
   }