+++ lib/screens/asignaturas_lista_screen.dart
@@ -1,11 +1,9 @@
-import 'package:extended_image/extended_image.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:mdi/mdi.dart';
 import 'package:mi_utem/config/routes.dart';
+import 'package:mi_utem/controllers/asignaturas_controller.dart';
 import 'package:mi_utem/models/asignatura.dart';
-import 'package:mi_utem/screens/asignatura_detalle_screen.dart';
-import 'package:mi_utem/services/asignaturas_service.dart';
 import 'package:mi_utem/services/config_service.dart';
 import 'package:mi_utem/themes/theme.dart';
 import 'package:mi_utem/widgets/custom_app_bar.dart';
@@ -13,34 +11,11 @@ import 'package:mi_utem/widgets/custom_error_widget.dart';
 import 'package:mi_utem/widgets/loading_indicator.dart';
 import 'package:mi_utem/widgets/pull_to_refresh.dart';
 
-class AsignaturasListaScreen extends StatefulWidget {
+class AsignaturasListaScreen extends GetView<AsignaturasController> {
   AsignaturasListaScreen({Key? key}) : super(key: key);
 
-  @override
-  _AsignaturasListaScreenState createState() => _AsignaturasListaScreenState();
-}
-
-class _AsignaturasListaScreenState extends State<AsignaturasListaScreen> {
-  Future<List<Asignatura>>? _futureAsignaturas;
-  late List<Asignatura> _asignaturas;
-
-  @override
-  void initState() {
-    super.initState();
-    _futureAsignaturas = _getAsignaturas();
-  }
-
-  Future<List<Asignatura>> _getAsignaturas([bool refresh = false]) async {
-    List<Asignatura> asignaturas =
-        await AsignaturasService.getAsignaturas(forceRefresh: refresh);
-    setState(() {
-      _asignaturas = asignaturas;
-    });
-    return asignaturas;
-  }
-
   Future<void> _onRefresh() async {
-    await _getAsignaturas(true);
+    controller.getAsignaturas();
   }
 
   bool get _mostrarCalculadora {
@@ -66,40 +41,12 @@ class _AsignaturasListaScreenState extends State<AsignaturasListaScreen> {
               ]
             : [],
       ),
-      body: FutureBuilder<List<Asignatura>>(
-        future: _futureAsignaturas,
-        builder: (context, snapshot) {
-          if (snapshot.hasError) {
-            return CustomErrorWidget(
-                title: "Ocurrió un error al obtener las asignaturas",
-                error: snapshot.error);
-          } else {
-            if (snapshot.hasData && snapshot.data != null) {
-              if (snapshot.data!.length > 0) {
-                return PullToRefresh(
-                  onRefresh: () async {
-                    await _onRefresh();
-                  },
-                  child: Padding(
-                    padding: const EdgeInsets.only(top: 8.0),
-                    child: ListView.builder(
-                      physics: AlwaysScrollableScrollPhysics(),
-                      itemBuilder: (BuildContext context, int i) {
-                        Asignatura asignatura = _asignaturas[i];
-                        return AsignaturaListTile(asignatura: asignatura);
-                      },
-                      itemCount: _asignaturas.length,
-                    ),
-                  ),
-                );
-              } else {
-                return CustomErrorWidget(
-                  emoji: "🤔",
-                  title: "Parece que no se encontraron asignaturas",
-                );
-              }
-            } else {
-              return Container(
+      body: PullToRefresh(
+        onRefresh: () async {
+          await _onRefresh();
+        },
+        child: Obx(() => controller.isLoading.value
+            ? Container(
                 padding: EdgeInsets.all(20),
                 child: Column(
                   mainAxisAlignment: MainAxisAlignment.center,
@@ -111,10 +58,20 @@ class _AsignaturasListaScreenState extends State<AsignaturasListaScreen> {
                     ),
                   ],
                 ),
-              );
-            }
-          }
-        },
+              )
+            : controller.asignaturas.isEmpty
+                ? CustomErrorWidget(
+                    emoji: "🤔",
+                    title: "Parece que no se encontraron asignaturas",
+                  )
+                : ListView.builder(
+                    physics: AlwaysScrollableScrollPhysics(),
+                    itemBuilder: (BuildContext context, int i) {
+                      Asignatura asignatura = controller.asignaturas[i];
+                      return AsignaturaListTile(asignatura: asignatura);
+                    },
+                    itemCount: controller.asignaturas.length,
+                  )),
       ),
     );
   }
@@ -138,12 +95,7 @@ class AsignaturaListTile extends StatelessWidget {
       ),
       child: Card(
         child: InkWell(
-          onTap: () => Get.to(
-            () => AsignaturaDetalleScreen(
-              asignatura: asignatura,
-            ),
-            routeName: Routes.asignatura,
-          ),
+          onTap: () => Get.toNamed('${Routes.asignatura}/${asignatura.id}'),
           child: Container(
             padding: EdgeInsets.all(20),
             width: double.infinity,