+++ CHANGELOG.md
@@ -26,7 +26,11 @@ Tipos de cambios
 - `Fixed`: para corrección de errores.
 - `Security`: en caso de vulnerabilidades.
 
-## [Unreleased]
+## [Unreleased] 
+
+### Added
+
+- Se añade controlador para Asignaturas
 
 ### Changed
 - Se cambió el estilo de la pantalla de notas/asignaturas. Ahora luce un poquito mejor.